Fix code::blocks compilation, using the C interface of FModEx finally.
git-svn-id: svn://solarus-engine.org/solarus@493 08e91b48-8f19-0410-8fc9-9db6b7bce602

+++ src/Savegame.cpp
@@ -61,8 +61,8 @@ void Savegame::set_initial_values(void) {
 
   set_integer(MAX_RUPEES, 99);
 
-  set_integer(ITEM_SLOT_0, -1);
-  set_integer(ITEM_SLOT_1, -1);
+  set_integer(ITEM_SLOT_0, (Uint32) -1);
+  set_integer(ITEM_SLOT_1, (Uint32) -1);
 }
 
 /**
